data-challenge/data_loader/CF/requirements.txt
pandas>=1.2.5
google-cloud-bigquery>=2.20.0
google-cloud-storage>=1.39.0